rename strace to trace
